Refactored key functions into utility files.
