+++ packages/varo/configuration-schema.json
@@ -1,36 +1,7 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "properties": {
-    "baseUrl": {
-      "title": "Base URL",
-      "anyOf": [
-        {
-          "type": "string"
-        },
-        {
-          "type": "null"
-        }
-      ],
-      "description": "The base URL (http://www.example.com)",
-      "format": "uri",
-      "minLength": 1,
-      "examples": ["https://example.com"]
-    },
-    "username": {
-      "title": "Username",
-      "type": "string",
-      "description": "Username",
-      "examples": ["test@openfn.org"]
-    },
-    "password": {
-      "title": "Password",
-      "type": "string",
-      "description": "Password",
-      "writeOnly": true,
-      "examples": ["@some(!)Str0ngp4ss0w0rd"]
-    }
-  },
+  "properties": {},
   "type": "object",
   "additionalProperties": true,
-  "required": ["password", "username"]
+  "required": []
 }
